requirements.txt
cffi==1.16.0
cryptography==42.0.7
dnspython==2.6.1
future==1.0.0
futures==3.0.5
ipwhois==1.2.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-whois==0.9.4
pytz==2024.1
six==1.16.0
zope.interface==6.4
